Move slideshow components
