+++ src/example-stories/guidedTour/GuidedTour.stories.tsx
