+++ src/example-stories/slideshow/inline.json
