+++ src/example-stories/slideshow/vna_custom.scss
