+++ src/ui-components/CanvasDetail/CanvasDetail.js
