+++ src/ui-components/CanvasDetail/CanvasDetail.scss
