+++ src/ui-components/CanvasNavigation/CanvasNavigation.scss
