+++ src/ui-components/Icons/Reset.tsx
