+++ src/ui-components/MobilePageView/MobilePageView.js
