+++ src/ui-components/MobilePageView/MobilePageView.scss
