+++ src/ui-components/MobileViewer/MobileViewer.scss
