+++ src/ui-components/PeekComponent/PeekComponent.js
