+++ src/ui-components/PeekComponent/PeekComponent.scss
