+++ src/ui-components/ProgressIndicator/ProgressIndicator.js
