+++ src/ui-components/ProgressIndicator/ProgressIndicator.scss
