+++ src/ui-components/SimpleSlideTransition/SimpleSlideTransition.js
