+++ src/ui-components/SimpleSlideTransition/SimpleSlideTransition.scss
