+++ src/ui-components/Slide/Slide.scss
