+++ src/ui-components/Slide/Slide.tsx
