+++ src/ui-components/Slideshow/SlideShow.scss
