+++ src/ui-components/SwappableViewer/SwappableViewer.scss
