+++ src/ui-components/TapDetector/TapDetector.js
