+++ src/ui-components/ZoomButtons/ZoomButtons.scss
@@ -1,22 +1,28 @@
 .zoom-buttons {
-  position: fixed;
-  top: 60px;
-  left: 20px;
+  position: absolute;
+  top: 14px;
+  left: 14px;
   z-index: 10000;
+
+  &--right {
+    left: auto;
+    right: 14px;
+  }
 
   &__button {
     background: transparent;
-    //width: 32px;
-    //height: 32px;
-    //background: rgb(255, 255, 255);
-    //box-shadow: rgba(0, 0, 0, 0.33) 0px 1px 3px 1px;
-    //border-radius: 50%;
-    //text-align: center;
     border: none;
     display: block;
     margin: 10px;
-    //font-size: 18px;
     outline: none;
+    cursor: pointer;
+    opacity: 1;
+    transition: opacity 0.5s;
+
+    &--disabled {
+      opacity: 0.5;
+      pointer-events: none;
+    }
 
     svg {
       transition: filter 0.3s;
