+++ src/utility/is-function.ts
