+++ src/utility/parse-manifest.ts
