+++ src/viewers/slideshow/index.js
@@ -1,16 +1,16 @@
-import CanvasDetail from './components/CanvasDetail/CanvasDetail';
-import CanvasNavigation from './components/CanvasNavigation/CanvasNavigation.tsx';
-import FullscreenButton from './components/FullscreenButton/FullscreenButton';
-import MobilePageView from './components/MobilePageView/MobilePageView';
-import MobileViewer from './components/MobileViewer/MobileViewer';
-import PeekComponent from './components/PeekComponent/PeekComponent';
-import ProgressIndicator from './components/ProgressIndicator/ProgressIndicator';
-import SimpleSlideTransition from './components/SimpleSlideTransition/SimpleSlideTransition';
-import Slide from './components/Slide/Slide.tsx';
-import SlideShow from './components/Slideshow/SlideShow.tsx';
-import SwappableViewer from './components/SwappableViewer/SwappableViewer.tsx';
-import TapDetector from './components/TapDetector/TapDetector';
-import ZoomButtons from './components/ZoomButtons/ZoomButtons';
+import CanvasDetail from '../../ui-components/CanvasDetail/CanvasDetail';
+import CanvasNavigation from '../../ui-components/CanvasNavigation/CanvasNavigation.tsx';
+import FullscreenButton from '../../ui-components/FullscreenButton/FullscreenButton';
+import MobilePageView from '../../ui-components/MobilePageView/MobilePageView';
+import MobileViewer from '../../ui-components/MobileViewer/MobileViewer';
+import PeekComponent from '../../ui-components/PeekComponent/PeekComponent';
+import ProgressIndicator from '../../ui-components/ProgressIndicator/ProgressIndicator';
+import SimpleSlideTransition from '../../ui-components/SimpleSlideTransition/SimpleSlideTransition';
+import Slide from '../../ui-components/Slide/Slide.tsx';
+import SlideShow from '../../ui-components/Slideshow/SlideShow.tsx';
+import SwappableViewer from '../../ui-components/SwappableViewer/SwappableViewer.tsx';
+import TapDetector from '../../ui-components/TapDetector/TapDetector';
+import ZoomButtons from '../../ui-components/ZoomButtons/ZoomButtons';
 
 const DEFAULT_SLIDESHOW_CONFIG = {
   mobileBreakpoint: 767,
