+++ src/viewers/slideshow/umd.js
